+++ game_patch/misc/alpine_settings.h
@@ -123,6 +123,7 @@ struct AlpineGameSettings
     bool always_clamp_official_lightmaps = false;
     bool static_bomb_code = false;
     bool entity_pain_sounds = true;
+    bool straight_armor_values = false;
 
     std::string multiplayer_tracker = "rfgt.factionfiles.com";
     static constexpr size_t max_tracker_hostname_length = 200;